Make a lib/ folder for lambdas
