+++ lambdas/tsconfig.json
@@ -1,6 +1,4 @@
 {
   "extends": "../tsconfig.json",
-  "include": [
-    "*.ts"
-  ]
+  "include": ["**/*.ts"]
 }
